requirements.txt
dbt-core
duckdb
dbt-duckdb
pandas
polars
sqlalchemy
pyodbc
faker
streamlit
pyarrow
snowflake-connector-python
pyyaml